+++ hellolog.php
@@ -3,7 +3,7 @@
  * Plugin Name:       helloLOG
  * Plugin URI:        https://hellowp.io/hellolog
  * Description:       Lightweight WordPress activity log. By hellowp.io and gobird.io.
- * Version:           0.3.1
+ * Version:           0.3.2
  * Requires at least: 6.4
  * Requires PHP:      8.0
  * Author:            HelloWP &amp; goBird
@@ -14,8 +14,7 @@
  * Domain Path:       /languages
  *
  * @package HelloLog
- * @version 0.3.1
- 
+ * @version 0.3.2
  */
 
 declare(strict_types=1);
@@ -52,7 +51,7 @@ if ( version_compare( PHP_VERSION, '8.0', '<' ) ) {
 	return;
 }
 
-const HELLOLOG_VERSION = '0.3.1';
+const HELLOLOG_VERSION = '0.3.2';
 const HELLOLOG_FILE    = __FILE__;
 
 define( 'HELLOLOG_DIR', plugin_dir_path( __FILE__ ) );